requirements.txt
bcrypt==4.2.0
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
convex==0.6.0
diskcache==5.6.3
filelock==3.15.4
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
fsspec==2024.6.1
huggingface-hub==0.24.6
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
llama_cpp_python==0.2.89
MarkupSafe==2.1.5
numpy==2.1.0
packaging==24.1
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.4
